add more functionality to the login buttons
change the functionality of the logout button on the header section of the page to make it disappear when the user is logged out and appear when the user is logged in.
Also change firebase.js to implement such functionality.
started the customization of user experience on the login buttons